fix(monitor): move SaludCard out of pages/ to bypass Vercel prerender bug
Files under src/pages/admin/monitor/ get registered as static routes by
Next.js Pages Router — including underscore-prefixed components like
_SaludCard. Local `next build` prerendered it fine to a benign loading
HTML, but Vercel's build machinery fails on the same prerender (3
attempts, opaque error). _CatalogTab/_MonitorTable/_ReviewBar happen
to survive because their loading paths don't touch the failing chunk,
but the pattern is fragile.

Move SaludCard to src/components/monitor/SaludCard.tsx so Next never
treats it as a page. Updates the import in [name].tsx.
