+++ src/components/monitor/SaludCard.tsx
